# Use zfs_arc_meta_limit_percent, zfs_arc_dnode_limit_percent properly