@dweeezil 2017-06-27
@behlendorf I'll take a look at this tomorrow.  IIRC, there were change changes made to the semantics of the dnode limiting tunable after the dnode limiting patch so I'll also want to review the history of those changes as well.  After a quick skim, however, it does look like a good idea to move the setting of ```arc_dnode_limit``` closer to the end of the tuning updates, after ```arc_meta_limit``` is dealt with.