@dweeezil 2017-06-29
@chrisrd I'm also fine with the changes as written.  If I have any issue at all, it would be that the "range comments" don't necessarily fully explain that the "additional" updates in each section are being done due to dependencies.  The issue being whether the situation would be clear to a developer in the future who's reading the code for the first time and, maybe, adding support for more tunables.